feat: add succes and error msgs
