+++ __tests__/errorMessage/errorMessage.test.ts
@@ -1,6 +1,6 @@
-import ErrorMsg from "../../src/app/services/error-msg/error-msg";
+import ToastMsg from "../../src/app/services/error-msg/toast-msg";
 
-const elementErrorMsg = new ErrorMsg();
+const elementErrorMsg = new ToastMsg();
 
 test("check type of result ErrorMsg class create", () => {
   expect(typeof elementErrorMsg).toBe("object");
