+++ src/app/services/api-request-service/api-request-service.ts
@@ -10,7 +10,7 @@ import { projectKey } from "./constants";
 import type StateManager from "../state-manager/state-manager";
 import type { Client, QueryParam, TokenStore } from "@commercetools/ts-client";
 import VSATokenCache from "./token-cache";
-import ErrorMsg from "../error-msg/error-msg";
+import ToastMsg from "../error-msg/toast-msg";
 import type { UseProductQuery, UseSearchQuery } from "../../global-types/types";
 
 type RequestBuilder = "anon" | "password";
@@ -43,12 +43,12 @@ export default class ApiRequestService {
   private stateManager: StateManager;
   private currentClientType!: RequestBuilder;
   private tokenCache: VSATokenCache;
-  private errorMsg: ErrorMsg;
+  private toastMsg: ToastMsg;
 
   constructor(stateManager: StateManager) {
     this.stateManager = stateManager;
     this.tokenCache = new VSATokenCache();
-    this.errorMsg = new ErrorMsg();
+    this.toastMsg = new ToastMsg();
     this.configureService();
   }
 
@@ -223,9 +223,11 @@ export default class ApiRequestService {
       })
       .execute()
       .then((result) => {
+        this.toastMsg.displaySuccessMsg(["Your data was successfully updated"]);
         if (onSuccess) onSuccess(result);
       })
       .catch((reason) => {
+        this.toastMsg.displayErrorMsg(ApiRequestService.errorParser(reason));
         if (onReject) onReject(reason);
       });
   }
@@ -305,10 +307,13 @@ export default class ApiRequestService {
       .execute()
       .then((result) => {
         this.switchRequestBuilder("anon");
+        this.toastMsg.displaySuccessMsg([
+          "Your password was successfully updated",
+        ]);
         if (onSuccess) onSuccess(result);
       })
       .catch((reason) => {
-        this.errorMsg.displayErrorMsg(ApiRequestService.errorParser(reason));
+        this.toastMsg.displayErrorMsg(ApiRequestService.errorParser(reason));
         if (onReject) onReject(reason);
       });
   }
