+++ src/app/services/error-msg/error-msg.module.css.d.ts
@@ -1,6 +1,8 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
+  'badMsg': string;
+  'goodMsg': string;
   'mainContainer': string;
   'myAnimation': string;
 }
